requirements.txt
Django==5.0.6
djangorestframework==3.15.1
qrcode==7.4.2
requests==2.31.0
gunicorn==21.2.0
django-cors-headers==4.4.0
